feat: refactors default layout and adds new component
